+++ src/components/AwesomeButtonProgress.vue
@@ -26,6 +26,7 @@ const props = withDefaults(defineProps<AwesomeButtonProgressProps>(), {
   disabled: false,
   visible: true,
   placeholder: false,
+  animateSize: true,
   textTransition: false,
   between: false,
   ripple: false,
@@ -231,6 +232,7 @@ function handlePress(event: ButtonPressEvent) {
     :disabled="props.disabled"
     :visible="props.visible"
     :placeholder="props.placeholder"
+    :animate-size="props.animateSize"
     :text-transition="props.textTransition"
     :between="props.between"
     :ripple="props.ripple"